requirements.txt
fastapi
uvicorn
pytesseract
pdfplumber
pillow
python-docx
reportlab
python-multipart